drivers: video: fbdev: exynos: dpu_9810: fence handling bug fix.
In kernel4.14, DPU driver fence handling method is changed to dma_fence.
But DPU driver did not increment the reference count of the fence for
dma_fence_signal.

So when the DPU driver decon_fence_signal, it was already approaching
the page that was released.

To solve this problem, when the sync_file->fence is linking to the
retire_fence of decon_reg_data. It's reference count is increased.
And release retire_fence code is added after retire_fence is signaled.

Change-Id: I148f26c082ce3229559c4e79a21dbaf35d188bf9
Signed-off-by: Jiun Yu <jiun.yu@samsung.com>
Signed-off-by: hwangjae lee <hj-yo.lee@samsung.com>